Removed pauses between iterations
After a discussion on the PR, I tested to see whether the pauses
which I had between iterations on some benchmarks made a notable
difference. After testing on odroidc2 MCS/non-MCS, which was the
platform that I observed these issues on which lead to me adding
the pause, I found that there was no significant difference
without the pause so I have removed them.

Signed-off-by: Alwin Joshy <joshyalwin@gmail.com>